Add Example of Method and Associated fun of struct
==============================

Add Example created in episode No 11.
In this file we have 2 structs: Course and Student.
We have one method and one associated function in impl block of the
Student to show multiple entity in a block.
Also we have two separate methods for the Cource struct in two separate
impl blocks to showing possibility of having more than one impl for a
struct.